created Directory.cs
uses namespace Tools, contains 2 methods for retrieving the current project directory on any machine, specifically a 'Temp' directory.
